added custom security expression

+++ src/main/java/ru/volga_it/simbir_go/features/account/AdminAccountController.java
@@ -1,5 +1,6 @@
 package ru.volga_it.simbir_go.features.account;
 
+import io.swagger.v3.oas.annotations.security.SecurityRequirement;
 import jakarta.validation.Valid;
 import jakarta.validation.constraints.Max;
 import jakarta.validation.constraints.Min;
@@ -9,6 +10,7 @@ import org.springframework.http.HttpStatus;
 import org.springframework.validation.annotation.Validated;
 import org.springframework.web.bind.annotation.*;
 import ru.volga_it.simbir_go.common.dto.PageDto;
+import ru.volga_it.simbir_go.common.exceptions.UnauthorizedException;
 import ru.volga_it.simbir_go.common.validation.OnCreate;
 import ru.volga_it.simbir_go.common.validation.OnUpdate;
 import ru.volga_it.simbir_go.features.account.dto.UserDto;
@@ -19,6 +21,7 @@ import ru.volga_it.simbir_go.features.account.mappers.UserMapper;
 import ru.volga_it.simbir_go.features.account.mappers.UserShortMapper;
 import ru.volga_it.simbir_go.features.account.mappers.admin.AdminCreateOrUpdateUserRequestBodyMapper;
 import ru.volga_it.simbir_go.features.account.services.UserService;
+import ru.volga_it.simbir_go.features.security.expressions.CustomSecurityExpression;
 
 @Valid
 @RestController
@@ -28,46 +31,63 @@ public class AdminAccountController {
 
     private final UserService userService;
 
+    private final CustomSecurityExpression customSecurityExpression;
+
     private final UserMapper userMapper;
     private final UserShortMapper userShortMapper;
     private final AdminCreateOrUpdateUserRequestBodyMapper adminCreateOrUpdateUserRequestBodyMapper;
 
     @GetMapping
     @ResponseStatus(HttpStatus.OK)
+    @SecurityRequirement(name = "bearerAuth")
     private PageDto<UserDto> getAll(
             @RequestParam(name = "start", defaultValue = "0") @Min(0) Integer offset,
             @RequestParam(name = "count", defaultValue = "20") @Min(1) @Max(100) Integer limit
     ) {
+        if(!customSecurityExpression.hasIsAdmin()) throw new UnauthorizedException("user is not an admin");
+
         Page<UserDto> page =  userService.getAll(offset, limit).map(userMapper::toDto);
         return new PageDto<>(page);
     }
 
     @GetMapping("{id}")
     @ResponseStatus(HttpStatus.OK)
+    @SecurityRequirement(name = "bearerAuth")
     private UserDto getById(@PathVariable Long id) {
+        if(!customSecurityExpression.hasIsAdmin()) throw new UnauthorizedException("user is not an admin");
+
         return userMapper.toDto(userService.getById(id));
     }
 
     @PostMapping
     @ResponseStatus(HttpStatus.CREATED)
+    @SecurityRequirement(name = "bearerAuth")
     private UserShortDto add(@Validated(OnCreate.class) @RequestBody AdminCreateOrUpdateUserRequestBody body) {
+        if(!customSecurityExpression.hasIsAdmin()) throw new UnauthorizedException("user is not an admin");
+
         UserEntity user = adminCreateOrUpdateUserRequestBodyMapper.toEntity(body);
         return userShortMapper.toDto(userService.add(user));
     }
 
     @PutMapping("{id}")
     @ResponseStatus(HttpStatus.NO_CONTENT)
+    @SecurityRequirement(name = "bearerAuth")
     private void update(
             @PathVariable Long id,
             @Validated(OnUpdate.class) @RequestBody AdminCreateOrUpdateUserRequestBody body
     ) {
+        if(!customSecurityExpression.hasIsAdmin()) throw new UnauthorizedException("user is not an admin");
+
         UserEntity user = adminCreateOrUpdateUserRequestBodyMapper.toEntity(body);
         userService.update(id, user);
     }
 
     @DeleteMapping("{id}")
     @ResponseStatus(HttpStatus.NO_CONTENT)
+    @SecurityRequirement(name = "bearerAuth")
     private void deleteById(@PathVariable Long id) {
+        if(!customSecurityExpression.hasIsAdmin()) throw new UnauthorizedException("user is not an admin");
+
         userService.deleteById(id);
     }
 }
