+++ src/main/java/ru/volga_it/simbir_go/features/transport/AdminTransportController.java
@@ -1,5 +1,6 @@
 package ru.volga_it.simbir_go.features.transport;
 
+import io.swagger.v3.oas.annotations.security.SecurityRequirement;
 import jakarta.validation.Valid;
 import jakarta.validation.constraints.Max;
 import jakarta.validation.constraints.Min;
@@ -9,7 +10,9 @@ import org.springframework.http.HttpStatus;
 import org.springframework.validation.annotation.Validated;
 import org.springframework.web.bind.annotation.*;
 import ru.volga_it.simbir_go.common.dto.PageDto;
+import ru.volga_it.simbir_go.common.exceptions.UnauthorizedException;
 import ru.volga_it.simbir_go.common.validation.OnUpdate;
+import ru.volga_it.simbir_go.features.security.expressions.CustomSecurityExpression;
 import ru.volga_it.simbir_go.features.transport.dto.TransportDetailsDto;
 import ru.volga_it.simbir_go.features.transport.dto.TransportDto;
 import ru.volga_it.simbir_go.features.transport.dto.params.TransportTypeRequestParam;
@@ -27,47 +30,64 @@ public class AdminTransportController {
 
     private final TransportService transportService;
 
+    private final CustomSecurityExpression customSecurityExpression;
+
     private final TransportMapper transportMapper;
     private final TransportDetailsMapper transportDetailsMapper;
 
     @GetMapping
     @ResponseStatus(HttpStatus.OK)
+    @SecurityRequirement(name = "bearerAuth")
     private PageDto<TransportDto> getAll(
             @RequestParam(name = "start", defaultValue = "0") @Min(0) Integer offset,
             @RequestParam(name = "count", defaultValue = "20") @Min(1) @Max(100) Integer limit,
             @RequestParam(defaultValue = "All") TransportTypeRequestParam type
     ) {
+        if(!customSecurityExpression.hasIsAdmin()) throw new UnauthorizedException("user is not an admin");
+
         Page<TransportDto> result = transportService.getAll(offset, limit, type).map(transportMapper::toDto);
         return new PageDto<>(result);
     }
 
     @GetMapping("{id}")
     @ResponseStatus(HttpStatus.OK)
+    @SecurityRequirement(name = "bearerAuth")
     private TransportDetailsDto getById(@PathVariable Long id) {
+        if(!customSecurityExpression.hasIsAdmin()) throw new UnauthorizedException("user is not an admin");
+
         return transportDetailsMapper.toDto(transportService.getById(id));
     }
 
     @PostMapping
     @ResponseStatus(HttpStatus.CREATED)
+    @SecurityRequirement(name = "bearerAuth")
     private TransportDto add(
             @Validated(OnUpdate.class) @RequestBody AdminCreateOrUpdateTransportParams params
     ) {
+        if(!customSecurityExpression.hasIsAdmin()) throw new UnauthorizedException("user is not an admin");
+
         TransportEntity transport = params.toEntity();
         return transportMapper.toDto(transportService.add(params.getOwnerId(), transport));
     }
 
     @PutMapping("{id}")
     @ResponseStatus(HttpStatus.NO_CONTENT)
+    @SecurityRequirement(name = "bearerAuth")
     private void update(
             @PathVariable Long id,
             @Validated(OnUpdate.class) @RequestBody AdminCreateOrUpdateTransportParams params
     ) {
+        if(!customSecurityExpression.hasIsAdmin()) throw new UnauthorizedException("user is not an admin");
+
         transportService.update(id, params);
     }
 
     @DeleteMapping("{id}")
     @ResponseStatus(HttpStatus.NO_CONTENT)
+    @SecurityRequirement(name = "bearerAuth")
     private void deleteById(@PathVariable Long id) {
+        if(!customSecurityExpression.hasIsAdmin()) throw new UnauthorizedException("user is not an admin");
+
         transportService.deletedById(id);
     }
 }
